TilemapRecolor/requirements.txt
Pillow>=9.0.0
numpy>=1.20.0
PyQt5>=5.15.0 